% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunShiny.R
\name{MERA}
\alias{MERA}
\title{Run the MERA Shiny App}
\usage{
MERA(skin = "Generic")
}
\arguments{
\item{skin}{Character. Skin to use to present results}
}
\value{
Nothing. Opens the MERA app in a web browser
}
\description{
Run the MERA Shiny App
}
